+++ src/main/java/com/ejemplo/spring/facturacion/bean/ComprobanteBean.java
@@ -6,14 +6,14 @@ public class ComprobanteBean implements Serializable
 {
 	private static final long serialVersionUID = 1L;
 	
-	private Integer ID;
-	private Integer numeroComprobante;
+	private Long ID;
+	private Double numeroComprobante;
 	private String nombreCliente;
 	private String apellidoPaterno;
 	private String apellidoMaterno;
 	private String DNI;
 	private String RUC;
-	private Float montototal;
+	private Double montototal;
 	private String estado;
 	private String sede;
 	
@@ -30,10 +30,10 @@ public class ComprobanteBean implements Serializable
 	public void setEstado(String estado) {
 		this.estado = estado;
 	}
-	public Integer getID() {
+	public Long getID() {
 		return ID;
 	}
-	public void setID(Integer iD) {
+	public void setID(Long iD) {
 		ID = iD;
 	}
 	public String getDNI() {
@@ -67,16 +67,16 @@ public class ComprobanteBean implements Serializable
 		this.apellidoMaterno = apellidoMaterno;
 	}
 
-	public Integer getNumeroComprobante() {
+	public Double getNumeroComprobante() {
 		return numeroComprobante;
 	}
-	public void setNumeroComprobante(Integer numeroComprobante) {
+	public void setNumeroComprobante(Double numeroComprobante) {
 		this.numeroComprobante = numeroComprobante;
 	}
-	public Float getMontototal() {
+	public Double getMontototal() {
 		return montototal;
 	}
-	public void setMontototal(Float montototal) {
+	public void setMontototal(Double montototal) {
 		this.montototal = montototal;
 	}
 	@Override
